+++ test/services/DummyCommandableCloudFunctionService_test.go
@@ -20,7 +20,7 @@ func newDummyCommandableAzureFunctionServiceTest() *DummyCommandableAzureFunctio
 func (c *DummyCommandableAzureFunctionServiceTest) setup(t *testing.T) {
 	config := cconf.NewConfigParamsFromTuples(
 		"logger.descriptor", "pip-services:logger:console:default:1.0",
-		"service.descriptor", "pip-services-dummies:service:commandable-azure-function:default:1.0",
+		"service.descriptor", "pip-services-dummies:service:commandable-azurefunc:default:1.0",
 	)
 
 	ctx := context.Background()